move phoca cart styling into its own module
To make it easier to maintain, move all of the rules related to
phoca cart into its own SCSS module.

The contents of phocacart.scss are sufficient to allow turning off
loading the main CSS file for this component. This is useful because
it clashes with the styling for the site.

This version is also slightly better at responsive sizing, especially
on the checkout screens.